:sparkles: Full modern revamping of n-todo-app
