+++ __test__/setup.ts
@@ -1,0 +1,2 @@
+import '@testing-library/preact';
+import 'preact/compat';
