+++ package.json
@@ -27,9 +27,13 @@
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
     "@preact/preset-vite": "^2.10.1",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/preact": "^3.2.4",
     "@types/node": "^24.0.1",
     "@types/react-transition-group": "^4.4.12",
+    "@vitest/coverage-v8": "3.2.3",
     "concurrently": "^9.1.2",
+    "happy-dom": "^18.0.1",
     "node": "^24.2.0",
     "prettier": "^3.5.3",
     "typed-css-modules": "^0.9.1",
@@ -39,8 +43,8 @@
     "vite-plugin-pwa": "^1.0.0",
     "vite-plugin-static-copy": "^3.0.0",
     "vite-plugin-svgr": "^4.3.0",
-    "vite-svg-loader": "^5.1.0",
-    "vite-tsconfig-paths": "^5.1.4"
+    "vite-tsconfig-paths": "^5.1.4",
+    "vitest": "^3.2.3"
   },
   "packageManager": "pnpm@10.12.1+sha512.f0dda8580f0ee9481c5c79a1d927b9164f2c478e90992ad268bbb2465a736984391d6333d2c327913578b2804af33474ca554ba29c04a8b13060a717675ae3ac"
 }
