+++ pnpm-workspace.yaml
@@ -1,3 +1,6 @@
+ignoredBuiltDependencies:
+  - node
+
 onlyBuiltDependencies:
   - '@biomejs/biome'
   - esbuild
