+++ src/components/index.ts
@@ -1,6 +1,6 @@
-export { default as AddButton } from './addButton';
-export { default as Button } from './button';
-export { default as Container } from './container';
+export { AddButton } from './AddButton';
+export { default as Button } from './Button/Button';
+export { Container } from './Container';
 export { default as Modal } from './modal';
 export { default as Input } from './input';
 export { default as Textarea } from './textarea';
