+++ src/main.tsx
@@ -2,7 +2,7 @@ import './global-style/normalize.css';
 import './global-style/index.css';
 import 'unfonts.css';
 import { render } from 'preact';
-import { App } from './app';
+import { App } from './App';
 import { APP_ROOT } from './globals';
 
 const root = document.getElementById(APP_ROOT);
