+++ src/vite-env.d.ts
@@ -2,6 +2,7 @@
 /// <reference types="vite-plugin-svgr/client" />
 /// <reference types="vite-plugin-pwa/client" />
 /// <reference types="vite-plugin-pwa/preact" />
+/// <reference types="vitest-dom/extend-expect" />
 
 declare module 'virtual:pwa-register/preact' {
   import type { StateUpdater } from 'preact/hooks';
