+++ tsconfig.app.json
@@ -13,7 +13,8 @@
       "@utils": ["./src/utils.ts"],
       "@components": ["./src/components/index.ts"],
       "@api/*": ["./src/api/*"],
-      "@hooks/*": ["./src/hooks/*"]
+      "@hooks/*": ["./src/hooks/*"],
+      "@globals": ["./src/globals.ts"]
     },
 
     /* Bundler mode */
